requirements.txt
flask
flask-sqlalchemy
flask-script
flask-migrate
werkzeug
pyjwt